move files into the /components
