+++ front/components/assistant_builder/AssistantBuilderDataSourceModal.tsx
@@ -14,8 +14,8 @@ import * as React from "react";
 import { useCallback, useEffect, useState } from "react";
 
 import { CONNECTOR_PROVIDER_TO_RESOURCE_NAME } from "@app/components/assistant_builder/AssistantBuilder";
+import type { AssistantBuilderDataSourceConfiguration } from "@app/components/assistant_builder/types";
 import DataSourceResourceSelectorTree from "@app/components/DataSourceResourceSelectorTree";
-import type { AssistantBuilderDataSourceConfiguration } from "@app/lib/api/assistant_builder/types";
 import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
 import type { GetConnectorResourceParentsResponseBody } from "@app/pages/api/w/[wId]/data_sources/[name]/managed/parents";
