+++ front/components/assistant_builder/AssistantBuilderDustAppModal.tsx
@@ -8,7 +8,7 @@ import {
 import type { AppType } from "@dust-tt/types";
 import { Transition } from "@headlessui/react";
 
-import type { AssistantBuilderDustAppConfiguration } from "@app/lib/api/assistant_builder/types";
+import type { AssistantBuilderDustAppConfiguration } from "@app/components/assistant_builder/types";
 
 export default function AssistantBuilderDustAppModal({
   isOpen,
