+++ front/components/assistant_builder/DataSourceSelectionSection.tsx
@@ -8,8 +8,8 @@ import {
 } from "@dust-tt/sparkle";
 
 import { CONNECTOR_PROVIDER_TO_RESOURCE_NAME } from "@app/components/assistant_builder/shared";
+import type { AssistantBuilderDataSourceConfiguration } from "@app/components/assistant_builder/types";
 import { EmptyCallToAction } from "@app/components/EmptyCallToAction";
-import type { AssistantBuilderDataSourceConfiguration } from "@app/lib/api/assistant_builder/types";
 import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
 import { getDisplayNameForDataSource } from "@app/lib/data_sources";
