+++ front/components/assistant_builder/server_side_props_helpers.ts
@@ -12,7 +12,7 @@ import {
   isTablesQueryConfiguration,
 } from "@dust-tt/types";
 
-import type { AssistantBuilderInitialState } from "@app/lib/api/assistant_builder/types";
+import type { AssistantBuilderInitialState } from "@app/components/assistant_builder/types";
 import { getDataSource } from "@app/lib/api/data_sources";
 import type { Authenticator } from "@app/lib/auth";
 import { tableKey } from "@app/lib/client/tables_query";
