+++ front/pages/w/[wId]/builder/assistants/new.tsx
@@ -18,13 +18,13 @@ import type { BuilderFlow } from "@app/components/assistant_builder/AssistantBui
 import AssistantBuilder, {
   BUILDER_FLOWS,
 } from "@app/components/assistant_builder/AssistantBuilder";
-import { getApps } from "@app/lib/api/app";
-import { getAgentConfiguration } from "@app/lib/api/assistant/configuration";
-import { buildInitialState } from "@app/lib/api/assistant_builder/server_side_helpers";
+import { buildInitialState } from "@app/components/assistant_builder/server_side_props_helpers";
 import type {
   AssistantBuilderDataSourceConfiguration,
   AssistantBuilderInitialState,
-} from "@app/lib/api/assistant_builder/types";
+} from "@app/components/assistant_builder/types";
+import { getApps } from "@app/lib/api/app";
+import { getAgentConfiguration } from "@app/lib/api/assistant/configuration";
 import { getDataSources } from "@app/lib/api/data_sources";
 import { Authenticator, getSession, getUserFromSession } from "@app/lib/auth";
